+++ src/sections/player.ts
@@ -1,18 +1,7 @@
-// debug logging.
-import Debug from 'debug/src/browser.js';
-import { DEBUG_APP_NAME } from '../constants';
-const debuglog = Debug(DEBUG_APP_NAME + ":player");
-
 // lovelace card imports.
-import { css, html, PropertyValues, TemplateResult } from 'lit';
+import { css, html, TemplateResult } from 'lit';
 import { customElement, property, state } from "lit/decorators.js";
 import { styleMap, StyleInfo } from 'lit-html/directives/style-map.js';
-
-// ** IMPORTANT - Vibrant notes:
-// ensure that you have "compilerOptions"."lib": [ ... , "WebWorker" ] specified
-// in your tsconfig.json!  If not, the Vibrant module will not initialize correctly
-// and you will tear your hair out trying to figure out why it doesn't work!!!
-import Vibrant from 'node-vibrant/dist/vibrant';
 
 // our imports - card components.
 import '../components/player-header';
@@ -32,9 +21,7 @@ import {
   PLAYER_CONTROLS_ICON_SIZE_DEFAULT
 } from '../constants';
 import { CardConfig } from '../types/card-config';
-import { Store } from '../model/store';
 import { MediaPlayer } from '../model/media-player';
-import { Palette } from '@vibrant/color';
 import { AlertUpdatesBase } from './alert-updates-base';
 
 
@@ -46,14 +33,6 @@ export class Player extends AlertUpdatesBase {
 
   // private storage.
   @state() private config!: CardConfig;
-  @state() private playerImage?: string;
-  @state() private _colorPaletteImage?: string;
-  @state() private _colorPaletteVibrant?: string;
-  @state() private _colorPaletteMuted?: string;
-  @state() private _colorPaletteDarkVibrant?: string;
-  @state() private _colorPaletteDarkMuted?: string;
-  @state() private _colorPaletteLightVibrant?: string;
-  @state() private _colorPaletteLightMuted?: string;
 
   /** MediaPlayer instance created from the configuration entity id. */
   private player!: MediaPlayer;
@@ -243,12 +222,14 @@ export class Player extends AlertUpdatesBase {
     const configImagePlayerOffBg = this.config.customImageUrls?.['playerOffBackground'];
     const configImageDefault = this.config.customImageUrls?.['default'];
 
-    // if player is off | idle | unknown, then reset the playerImage value so that one
-    // of the default images is selected below.
-    if (this.player.isPoweredOffOrIdle()) {
-      this.playerImage = undefined;
-      if (!this.config.playerMinimizeOnIdle) {
-        this.store.card.footerBackgroundColor = undefined;
+    // get current media player image and media content id values.
+    let playerMediaContentId:string | undefined = undefined;
+    let playerImage:string | undefined = undefined;
+    if (this.store.player) {
+      playerMediaContentId = this.store.player.attributes.media_content_id;
+      playerImage = (this.store.player.attributes.entity_picture || this.store.player.attributes.entity_picture_local);
+      if (playerImage) {
+        playerImage = this.store.hass.hassUrl(playerImage);
       }
     }
 
@@ -259,7 +240,7 @@ export class Player extends AlertUpdatesBase {
     //  "color:red",
     //  JSON.stringify(isOff),
     //  JSON.stringify(isIdle),
-    //  JSON.stringify(this.playerImage),
+    //  JSON.stringify(playerImage),
     //  JSON.stringify(this.config.playerMinimizeOnIdle),
     //  JSON.stringify(configImagePlayerIdleBg),
     //  JSON.stringify(configImagePlayerOffBg),
@@ -275,6 +256,7 @@ export class Player extends AlertUpdatesBase {
     if (isIdle && configImagePlayerIdleBg) {
 
       // use configured player idle background image.
+      this.store.card.playerMediaContentId = "configImagePlayerIdleBg"
       imageUrl = configImagePlayerIdleBg;
       if ((imageUrl + "").toLowerCase() == "none") {
         imageUrl = "";
@@ -286,6 +268,7 @@ export class Player extends AlertUpdatesBase {
     } else if (isOff && configImagePlayerOffBg) {
 
       // use configured player off background image.
+      this.store.card.playerMediaContentId = "configImagePlayerOffBg"
       imageUrl = configImagePlayerOffBg;
       if ((imageUrl + "").toLowerCase() == "none") {
         imageUrl = "";
@@ -297,30 +280,39 @@ export class Player extends AlertUpdatesBase {
     } else if (configImagePlayerBg) {
 
       // use configured player background image (static image, does not change).
+      this.store.card.playerMediaContentId = "configImagePlayerBg"
       imageUrl = configImagePlayerBg;
       headerBackgroundColor = this.config.playerHeaderBackgroundColor || PLAYER_CONTROLS_BACKGROUND_COLOR_DEFAULT;
       controlsBackgroundColor = this.config.playerControlsBackgroundColor || PLAYER_CONTROLS_BACKGROUND_COLOR_DEFAULT;
 
-    } else if (this.playerImage) {
+    } else if (playerImage) {
 
       // use currently playing artwork background image; image changes with the track.
-      imageUrl = this.playerImage;
+      this.store.card.playerMediaContentId = playerMediaContentId;
+      imageUrl = playerImage || "";
       headerBackgroundColor = this.config.playerHeaderBackgroundColor || PLAYER_CONTROLS_BACKGROUND_COLOR_DEFAULT;
       controlsBackgroundColor = this.config.playerControlsBackgroundColor || PLAYER_CONTROLS_BACKGROUND_COLOR_DEFAULT;
 
     } else if (configImageDefault) {
 
       // use configured default background image.
+      this.store.card.playerMediaContentId = "configImageDefault"
       imageUrl = configImageDefault;
       backgroundSize = BRAND_LOGO_IMAGE_SIZE;
-      this.store.card.footerBackgroundColor = undefined;
 
     } else {
 
-      // use hard-coded default background image.
-      imageUrl = BRAND_LOGO_IMAGE_BASE64;
-      backgroundSize = BRAND_LOGO_IMAGE_SIZE;
-      this.store.card.footerBackgroundColor = undefined;
+      this.store.card.playerMediaContentId = "BRAND_LOGO_IMAGE_BASE64"
+
+      // if minimized and we are idle | off then do not use the brand logo.
+      if (this.config.playerMinimizeOnIdle && (isIdle || isOff)) {
+        imageUrl = "";
+        backgroundSize = undefined;
+      } else {
+        // use hard-coded default background image.
+        imageUrl = BRAND_LOGO_IMAGE_BASE64;
+        backgroundSize = BRAND_LOGO_IMAGE_SIZE;
+      }
 
     }
 
@@ -344,6 +336,7 @@ export class Player extends AlertUpdatesBase {
 
     // build style info object.
     styleInfo['background-image'] = `url(${imageUrl})`;
+    this.store.card.playerImage = imageUrl;
     if (backgroundSize)
       styleInfo['--spc-player-background-size'] = `${backgroundSize}`;
     styleInfo['--spc-player-header-bg-color'] = `${headerBackgroundColor}`;
@@ -381,13 +374,6 @@ export class Player extends AlertUpdatesBase {
       styleInfo['--spc-player-volume-label-color'] = `${playerVolumeLabelColor}`;
     if (playerVolumeSliderColor)
       styleInfo['--spc-player-volume-slider-color'] = `${playerVolumeSliderColor}`;
-
-    styleInfo['--spc-player-palette-vibrant'] = `${this._colorPaletteVibrant}`;
-    styleInfo['--spc-player-palette-muted'] = `${this._colorPaletteMuted}`;
-    styleInfo['--spc-player-palette-darkvibrant'] = `${this._colorPaletteDarkVibrant}`;
-    styleInfo['--spc-player-palette-darkmuted'] = `${this._colorPaletteDarkMuted}`;
-    styleInfo['--spc-player-palette-lightvibrant'] = `${this._colorPaletteLightVibrant}`;
-    styleInfo['--spc-player-palette-lightmuted'] = `${this._colorPaletteLightMuted}`;
 
     return styleMap(styleInfo);
 
@@ -439,200 +425,4 @@ export class Player extends AlertUpdatesBase {
     return styleMap(styleInfo);
   }
 
-
-  /**
-   * Invoked before `update()` to compute values needed during the update.
-   * 
-   * We will check for changes in the media player background image.  If a
-   * change is being made, then we will analyze the new image for the vibrant
-   * color palette.  We will then set some css variables with those values for
-   * use by the different player sections (header, progress, volume, etc). 
-   */
-  protected willUpdate(changedProperties: PropertyValues): void {
-
-    // invoke base class method.
-    super.willUpdate(changedProperties);
-
-    // get list of changed property keys.
-    const changedPropKeys = Array.from(changedProperties.keys())
-
-    //if (debuglog.enabled) {
-    //  debuglog("willUpdate - changed property keys:\n%s\n- PlayerImage = %s",
-    //    JSON.stringify(changedPropKeys),
-    //    JSON.stringify(this.playerImage),
-    //  );
-    //}
-
-    // we only care about "store" property changes at this time, as it contains a
-    // reference to the "hass" property.  we are looking for background image changes.
-    if (!changedPropKeys.includes('store')) {
-      return;
-    }
-
-    let oldImage: string | undefined = undefined;
-    let newImage: string | undefined = undefined;
-    let oldMediaContentId: string | undefined = undefined;
-    let newMediaContentId: string | undefined = undefined;
-
-    // get the old property reference.
-    const oldStore = changedProperties.get('store') as Store;
-    if (oldStore) {
-
-      // if a media player was assigned to the store, then get the player image.
-      // convert the image url from a relative value to a fully-qualified url.
-      const oldPlayer = oldStore.player;
-      if (oldPlayer) {
-        oldImage = (oldPlayer.attributes.entity_picture || oldPlayer.attributes.entity_picture_local);
-        if (oldImage) {
-          oldImage = this.store.hass.hassUrl(oldImage);
-          oldMediaContentId = oldPlayer.attributes.media_content_id;
-        }
-      }
-    }
-
-    // check if the player reference is set (in case it was set to undefined).
-    if (this.store.player) {
-
-      // get the current media player image.
-      // if image not set, then there's nothing left to do.
-      newImage = (this.store.player.attributes.entity_picture || this.store.player.attributes.entity_picture_local);
-      if (newImage) {
-        newImage = this.store.hass.hassUrl(newImage);
-        newMediaContentId = this.store.player.attributes.media_content_id;
-        this.playerImage = newImage;
-      } else {
-        return;
-      }
-    }
-
-    //console.log("willUpdate - player content values:\n- OLD CONTENT ID = %s\n- NEW CONTENT ID = %s\n- mediaContentId = %s\n- OLD IMAGE = %s\n- NEW IMAGE = %s\n- PlayerImage = %s\n- footerBackgroundColor = %s",
-    //  JSON.stringify(oldMediaContentId),
-    //  JSON.stringify(newMediaContentId),
-    //  JSON.stringify(this.mediaContentId),
-    //  JSON.stringify(oldImage),
-    //  JSON.stringify(newImage),
-    //  JSON.stringify(this.playerImage),
-    //  JSON.stringify(this.store.card.footerBackgroundColor),
-    //);
-
-    // only need to update vibrant colors if we are not editing the card.
-    if (!this.isCardInEditPreview) {
-
-      // did the content change? 
-      // if so, then extract the color differences from the associated image.
-      // note that we cannot compare images here, as it's a cached value and the `cache` portion of
-      // image url could change even though it's the same content that's playing!
-      if (oldMediaContentId != newMediaContentId) {
-
-        if (debuglog.enabled) {
-          debuglog("willUpdate - player content changed:\n- OLD CONTENT ID = %s\n- NEW CONTENT ID = %s\n- mediaContentId = %s\n- OLD IMAGE = %s\n- NEW IMAGE = %s",
-            JSON.stringify(oldMediaContentId),
-            JSON.stringify(newMediaContentId),
-            JSON.stringify(this.mediaContentId),
-            JSON.stringify(oldImage),
-            JSON.stringify(newImage),
-          );
-        }
-
-        // extract the color differences from the new image and set the player colors.
-        this._extractColors();
-
-        // store the new media id in the exposed property so that other forms
-        // are informed of the change.
-        this.mediaContentId = newMediaContentId || "";
-
-      }
-    }
-  }
-
-
-  /**
-   * Extracts color compositions from the background image, which will be used for 
-   * rendering controls that are displayed on top of the background image.
-   * 
-   * Good resource on the Vibrant package parameters, examples, and other info:
-   * https://github.com/Vibrant-Colors/node-vibrant
-   * https://kiko.io/post/Get-and-use-a-dominant-color-that-matches-the-header-image/
-   * https://jariz.github.io/vibrant.js/
-   * https://github.com/Vibrant-Colors/node-vibrant/issues/44
-   */
-  private async _extractColors(): Promise<void> {
-
-    //console.log("_extractColors (player) - extracting vibrant colors from image:\n- playerImage = %s",
-    //  JSON.stringify(this.playerImage),
-    //);
-
-    //console.log("_extractColors (player) - colors before extract:\n- Vibrant      = %s\n- Muted        = %s\n- DarkVibrant  = %s\n- DarkMuted    = %s\n- LightVibrant = %s\n- LightMuted   = %s",
-    //  this._colorPaletteVibrant,
-    //  this._colorPaletteMuted,
-    //  this._colorPaletteDarkVibrant,
-    //  this._colorPaletteDarkMuted,
-    //  this._colorPaletteLightVibrant,
-    //  this._colorPaletteLightMuted,
-    //);
-
-    if (this.playerImage) {
-
-      // set options for vibrant call.
-      const vibrantOptions = {
-        "colorCount": 64, // amount of colors in initial palette from which the swatches will be generated.
-        "quality": 3,     // quality. 0 is highest, but takes way more processing.
-      //  "quantizer": 'mmcq',
-      //  "generators": ['default'],
-      //  "filters": ['default'],
-      }
-
-      // create vibrant instance with our desired options.
-      const vibrant: Vibrant = new Vibrant(this.playerImage || '', vibrantOptions);
-      this._colorPaletteImage = this.playerImage || '';
-      if (this._colorPaletteImage) { }  // keep the compiler happy!
-
-      // get the color palettes for the player background image.
-      await vibrant.getPalette().then(
-        (palette: Palette) => {
-
-          //console.log("_extractColors (player) - colors found by getPalette:\n- Vibrant      = %s\n- Muted        = %s\n- DarkVibrant  = %s\n- DarkMuted    = %s\n- LightVibrant = %s\n- LightMuted   = %s",
-          //  (palette['Vibrant']?.hex) || 'undefined',
-          //  (palette['Muted']?.hex) || 'undefined',
-          //  (palette['DarkVibrant']?.hex) || 'undefined',
-          //  (palette['DarkMuted']?.hex) || 'undefined',
-          //  (palette['LightVibrant']?.hex) || 'undefined',
-          //  (palette['LightMuted']?.hex) || 'undefined',
-          //);
-
-          // set player color palette values.
-          this._colorPaletteVibrant = (palette['Vibrant']?.hex) || undefined;
-          this._colorPaletteMuted = (palette['Muted']?.hex) || undefined;
-          this._colorPaletteDarkVibrant = (palette['DarkVibrant']?.hex) || undefined;
-          this._colorPaletteDarkMuted = (palette['DarkMuted']?.hex) || undefined;
-          this._colorPaletteLightVibrant = (palette['LightVibrant']?.hex) || undefined;
-          this._colorPaletteLightMuted = (palette['LightMuted']?.hex) || undefined;
-
-          // set card footer background color.
-          this.store.card.footerBackgroundColor = this._colorPaletteVibrant;
-
-        },
-        (_reason: string) => {
-
-          if (debuglog.enabled) {
-            debuglog("_extractColors - Could not retrieve color palette info for player background image\nreason = %s",
-              JSON.stringify(_reason),
-            );
-          }
-
-          // reset player color palette values.
-          this._colorPaletteVibrant = undefined;
-          this._colorPaletteMuted = undefined;
-          this._colorPaletteDarkVibrant = undefined;
-          this._colorPaletteDarkMuted = undefined;
-          this._colorPaletteLightVibrant = undefined;
-          this._colorPaletteLightMuted = undefined;
-
-          // set card footer background color.
-          this.store.card.footerBackgroundColor = this._colorPaletteVibrant;
-
-        }
-      );
-    }
-  }
 }